// src/components/JCachePage/typings.d.ts
import { useAppData } from '@umijs/max';
import { RouteMatch, RouteObject } from 'react-router-dom';
import { Location, Action } from 'history';

interface Routes
  extends otherRouteParams,
    Pick<ReturnType<typeof useAppData>, 'routes'> {
  /**
   * path 只支持两种占位符配置，
   * 第一种是动态参数 :id 的形式，
   * 第二种是 * 通配符，通配符只能出现路由字符串的最后。
   */
  path: string;
  /**
   * 配置 location 和 path 匹配后用于渲染的 React 组件路径。
   * 可以是绝对路径，也可以是相对路径，如果是相对路径，会从 src/pages 开始寻找。
   * 如果指向 src 目录的文件，可以用 @，
   * 比如 component: '@/layouts/basic'，推荐使用 @ 组织路由文件位置。
   */
  component?: string;
  // /**
  //  * 配置子路由，通常在需要为多个路径增加 layout 组件时使用
  //  */
  // routes?: Routes[];
  /**
   * 配置路由的标题
   */
  title?: string;
  /**
   * 重定向
   * 配置路由跳转。
   */
  redirect?: string; //
  /**
   * 配置路由组件的包装组件，通过包装组件可以为当前的路由组件组合进更多的功能。
   * 比如，可以用于路由级别的权限校验
   */
  wrapper?: string[];
}

interface RouteProps extends RouteMatch<string> {
  route: RouteObject & {
    title?: string;
  } & otherRouteParams;
  location: StackLocation;
}

type otherRouteParams = {
  // 以下为新增
  /**
   * 是否缓存当前界面，跳转到下一个路由不会卸载当前路由组件
   */
  keepAlive?: boolean;
  /**
   * 该界面是否自定义header，如果自定义，需要开发自己在自己的组件上使用Header
   */
  customHeader?: boolean;
};
type HistoryRouteProps = {
  back?: () => void; // 可以控制每一个界面的回退操作
  header?: React.ReactNode; // 自定义头部
};
type HistoryState =
  | {
      props: HistoryRouteProps;
      [key: string]: any;
    }
  | (string | boolean | number);
interface StackLocation extends Location {
  state: HistoryState;
  action: Action;
}
type RouteStack = Partial<RouteProps> & {
  location: StackLocation | Location;
};

interface CachePageType {
  routes: Partial<Routes>;
  currentRoute: RouteProps | null;
  routeStack: RouteProps[];
  routeProps: Record<string, HistoryRouteProps>;
}

type ChangePayloadType = {
  type: string;
  payload: Partial<CachePageType>;
};